+++ internal/listing/dirsize_other_test.go
@@ -1,0 +1,9 @@
+//go:build !unix
+
+package listing
+
+import "os"
+
+func sparseFixtureDiskUsage(info os.FileInfo) (int64, bool) {
+	return 0, false
+}